Persist multiple line weights and colors for different nibs
Nib points, color, line weight, and tip (sphere) are now stored in the nib_objects JSON. Works well, but a fair bit of added complication re: logic of updating color + line weight, as well as how to reset the scene, and even the undo button (if the most recent part is a nib).

But it works (for now...).